back/requirements.txt
opencv-python
pillow
numpy 
joblib
uvicorn
fastapi
